+++ src/components/common/SearchResultSummary.tsx
